Adding option to filter which variables should be loaded by the reader
This is expected to be used mostly when more than one (u,v) pairs are available in a file, to avoid  (re)mapping u or v components (e.g. tidal, residual or currents) to the same CF  names :  'x_sea_water_velocity' or 'y_sea_water_velocity'.
This can be also used when we want to read only a single variable from a file (e.g. bathymetry/depth); added some modifications to make sure than the bathymetry can be used even it the "time" associated with it in the file is not within the opendrift simulation time.