Add habitat settlement, vertical swimming, and Haliotis iris behavior
Add settlement restricted to suitable habitat only. The habitat is specified with an input shapefile and the code automatically extract the habitat polygons, compute their centroids, and organizes them in a Ball Tree for fast reference. When larvae are competent to settle, the model looks whether they are on top of a suitable habitat. If it is the case, the larvae can settle and their final position is recorded.

 Add vertical swimming behavior when larvae are being advected away from their habitat. When larvae detect that they are being advected away from the preferred habitat, they start to swim up and down using a correlated random walk (use of a persistence parameter) and stop when they find a more favorable current that bring them back to shore. This vertical sampling behavior is theoretical.

Add maximum depth of dispersal to limit the dispersal of larvae to depths where they have been observed and sampled.

Add Haliotis iris specific behavior: larvae sink during the first 12 hours, then swim upward.

See example_bivalve_larvae_Haliotis_iris.py for an example script to run the model.