reader_netCDF_CF_generic: bugfix for lazy loading with coordinates in decreasing order
The part of the grid corresponding to the active particles is calculated using xmin and ymin, but this should be x[0] and y[0]. For coordinates in decreasing order the minimum is not the same as the first value and the wrong part of the grid would be loaded. The use of np.abs here avoids having to use a signed delta_x and delta_y, which could break code in other places.