Allow for saving of interpolators
I made changes to structured.py to allow for saving the interpolators to save time on repeated simulations. The necessary attributes to use the save option need to be input via the reader, currently only available in the ROMS reader.

This PR also has changes that clean up the name variable in the ROMS reader. Previously there was logic to deal with name for several cases but subsequently it was overwritten to be “roms native”. Now name is used if input and otherwise set to “roms native”.