Typo fix to connect ROMS reader to save interpolator
I had the interpolator save and reuse working but did not realize it was not using the path I entered through
the ROMS reader. I realized now that I called the interpolator filename variable two different things! I have
fixed this now and also modified a few logger messages to be clearer.