Possible bug in LCS computation ?
It seems that opendrift inverts the order of lon,lat when running in backwards mode. 

So we'd need to flip the arrays of final (lon,lat) (when doing backwards simulation for the ALCS) to correctly compute the displacements that are then fed to method ftle(). 

This suppresses the need to flip the array after the ftle() call.